requirements.txt
pyyaml
requests
fastapi
uvicorn[standard]
pydantic